Deployment/requirements.txt
streamlit
pandas
seaborn
matplotlib
numpy
plotly
Pillow
scikit-learn==1.3.0
xgboost==2.0.3